refactor: Dynamic Proxy 활용하여 동적으로 트랜잭션 처리
