+++ app/src/main/java/com/techcourse/service/TransactionHandler.java
@@ -1,0 +1,47 @@
+package com.techcourse.service;
+
+import org.springframework.dao.DataAccessException;
+import org.springframework.jdbc.datasource.DataSourceUtils;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
+
+import javax.sql.DataSource;
+import java.lang.reflect.InvocationHandler;
+import java.lang.reflect.Method;
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public class TransactionHandler implements InvocationHandler {
+
+    private final Object service;
+    private final DataSource dataSource;
+
+    public TransactionHandler(Object service, DataSource dataSource) {
+        this.service = service;
+        this.dataSource = dataSource;
+    }
+
+    @Override
+    public Object invoke(Object proxy, Method method, Object[] args) throws Throwable {
+        Connection connection = DataSourceUtils.getConnection(dataSource);
+        TransactionSynchronizationManager.bindResource(dataSource, connection);
+        try {
+            connection.setAutoCommit(false);
+            Object invoke = method.invoke(service, args);
+            connection.commit();
+            return invoke;
+        } catch (Exception e) {
+            rollback(connection);
+            throw new DataAccessException(e);
+        } finally {
+            DataSourceUtils.releaseConnection(connection, dataSource);
+        }
+    }
+
+    private void rollback(Connection connection) {
+        try {
+            connection.rollback();
+        } catch (SQLException ex) {
+            throw new DataAccessException(ex);
+        }
+    }
+}
