--- app/src/main/java/com/techcourse/service/TxUserService.java
@@ -1,76 +1,0 @@
-package com.techcourse.service;
-
-import com.techcourse.domain.User;
-import org.springframework.dao.DataAccessException;
-import org.springframework.jdbc.datasource.DataSourceUtils;
-import org.springframework.transaction.support.TransactionSynchronizationManager;
-
-import javax.sql.DataSource;
-import java.sql.Connection;
-import java.sql.SQLException;
-
-public class TxUserService implements UserService {
-
-    private final UserService userService;
-    private final DataSource dataSource;
-
-    public TxUserService(UserService userService, DataSource dataSource) {
-        this.userService = userService;
-        this.dataSource = dataSource;
-    }
-
-    public User findById(final long id) {
-        Connection connection = DataSourceUtils.getConnection(dataSource);
-        TransactionSynchronizationManager.bindResource(dataSource, connection);
-        try {
-            connection.setAutoCommit(false);
-            User user = userService.findById(id);
-            connection.commit();
-            return user;
-        } catch (Exception e) {
-            rollback(connection);
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
-    }
-
-
-    public void insert(final User user) {
-        Connection connection = DataSourceUtils.getConnection(dataSource);
-        TransactionSynchronizationManager.bindResource(dataSource, connection);
-        try {
-            connection.setAutoCommit(false);
-            userService.insert(user);
-            connection.commit();
-        } catch (Exception e) {
-            rollback(connection);
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
-    }
-
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        Connection connection = DataSourceUtils.getConnection(dataSource);
-        TransactionSynchronizationManager.bindResource(dataSource, connection);
-        try {
-            connection.setAutoCommit(false);
-            userService.changePassword(id, newPassword, createBy);
-            connection.commit();
-        } catch (Exception e) {
-            rollback(connection);
-            throw new DataAccessException(e);
-        } finally {
-            DataSourceUtils.releaseConnection(connection, dataSource);
-        }
-    }
-
-    private void rollback(Connection connection) {
-        try {
-            connection.rollback();
-        } catch (SQLException ex) {
-            throw new DataAccessException(ex);
-        }
-    }
-}
